Initial Eclipse OMR web site implementation
Designed to be built as a static web site with Jekyll. I have not included
some not strictly necessary third party components from this version of
the web site so that we can start working on the other parts of our site
infrastructure (i.e. the job to build the site) while we evaluate the best
way to bring them in (things like Prism.js, Retina.js, DropIt.js).

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>